+++ cloud-yblog-article/cloud-yblog-article2801/src/main/java/com/boot/service/ArticleService.java
@@ -11,7 +11,7 @@ public interface ArticleService {
 
     public List<Article> selectAllArticle();
 
-    public Article selectArticleByArticleIdNoComment(Integer id);
+    public Article selectArticleByArticleIdNoComment(long id);
 
     //排行榜
     public List<Article> selectAllArticleOrderByDesc();
@@ -26,10 +26,10 @@ public interface ArticleService {
 
     public int addArticle(Article article);
 
-    public int deleteArticleByArticleId(Integer id);
+    public int deleteArticleByArticleId(long id);
 
     //当用户点击文章时，文章点击数加1
-    public void updateHits(Integer article_id);
+    public void updateHits(long article_id);
 
     public List<Article> selectCategoriesAndTags();
 
@@ -37,15 +37,15 @@ public interface ArticleService {
 
     public String selectTagsByArticleId(long id);
 
-    public void updateTagsByArticleId(String tags, int id);
+    public void updateTagsByArticleId(String tags, long id);
 
     public void updateCategory(String oldName, String newName);
 
     //是否可以评论功能
     //变成0
-    void updateAllowCommentTo_0(int id);
+    void updateAllowCommentTo_0(long id);
     //变成1
-    void updateAllowCommentTo_1(int id);
+    void updateAllowCommentTo_1(long id);
 
 
     /**
@@ -56,15 +56,15 @@ public interface ArticleService {
 
     void publishArticle_service(Article article);
 
-    void deleteArticle_service(Integer id);
+    void deleteArticle_service(long id);
 
     //echarts
     List<Article> selectArticleStatistic();
 
     //修改是否推荐
-    void updateRecommendTo_0(int id);
+    void updateRecommendTo_0(long id);
 
-    void updateRecommendTo_1(int id);
+    void updateRecommendTo_1(long id);
 
     //查询推荐的文章
     List<Article> selectArticleByRecommend();
