+++ cloud-yblog-article/cloud-yblog-article2801/src/main/java/com/boot/service/impl/LinkServiceImpl.java
@@ -30,7 +30,7 @@ public class LinkServiceImpl implements LinkService {
     }
 
     @Override
-    public void deleteLink(int id) {
+    public void deleteLink(long id) {
         linkMapper.deleteLink(id);
     }
