+++ cloud-yblog-article/cloud-yblog-article2801/src/main/java/com/boot/service/impl/StatisticServiceImpl.java
@@ -19,7 +19,7 @@ public class StatisticServiceImpl implements StatisticService {
     }
 
     @Override
-    public int deleteStatisticByArticleId(Integer id) {
+    public int deleteStatisticByArticleId(long id) {
         return statisticMapper.deleteStatisticByArticleId(id);
     }
 }
